requirements.txt
mistralai==1.7.1
streamlit==1.45.1
python-dotenv==1.1.0